# docs/internal_swagger.yaml
definitions:
  ErrorStruct:
    properties:
      error_code:
        type: integer
      error_message:
        type: string
    type: object
  domain.City:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      region_id:
        type: string
      updated_at:
        type: string
    type: object
  domain.GroupType:
    enum:
    - pensioners
    - disabled
    - young_families
    - low_income
    - students
    - large_families
    - children
    - veterans
    type: string
    x-enum-varnames:
    - UserGroupPensioners
    - UserGroupDisabled
    - UserGroupYoungFamilies
    - UserGroupLowIncome
    - UserGroupStudents
    - UserGroupLargeFamilies
    - UserGroupChildren
    - UserGroupVeterans
  domain.UserGroup:
    properties:
      error_message:
        description: Сообщение об ошибке
        type: string
      expires_at:
        description: Когда истекает
        type: string
      external_id:
        description: ID из внешнего API
        type: string
      rejected_at:
        description: Когда отклонена
        type: string
      status:
        $ref: '#/definitions/domain.VerificationStatus'
      type:
        $ref: '#/definitions/domain.GroupType'
      verified_at:
        description: Когда подтверждена
        type: string
    type: object
  domain.VerificationStatus:
    enum:
    - pending
    - verified
    - rejected
    - expired
    type: string
    x-enum-comments:
      VerificationStatusExpired: Истекла
      VerificationStatusPending: Ожидает подтверждения
      VerificationStatusRejected: Отклонена
      VerificationStatusVerified: Подтверждена
    x-enum-varnames:
    - VerificationStatusPending
    - VerificationStatusVerified
    - VerificationStatusRejected
    - VerificationStatusExpired
  v1.benefitResponse:
    properties:
      category:
        type: string
      city_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      how_to_use:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      region:
        items:
          type: integer
        type: array
      requirement:
        type: string
      source_url:
        type: string
      tags:
        items:
          type: string
        type: array
      target_groups:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
      views:
        type: integer
    type: object
  v1.benefitsListResponse:
    properties:
      benefits:
        items:
          $ref: '#/definitions/v1.benefitResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  v1.exchangeTokenRequest:
    properties:
      code:
        type: string
      state:
        type: string
    required:
    - code
    - state
    type: object
  v1.exchangeTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  v1.getProfileResponse:
    properties:
      city_id:
        type: string
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      groups:
        items:
          $ref: '#/definitions/domain.UserGroup'
        type: array
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      registered_at:
        type: string
      snils:
        type: string
    type: object
  v1.speechResponse:
    properties:
      text:
        type: string
    type: object
  v1.userUpdateInfoRequest:
    properties:
      city_id:
        type: string
      groups:
        items:
          $ref: '#/definitions/domain.GroupType'
        type: array
    required:
    - city_id
    - groups
    type: object
info:
  contact: {}
paths:
  /benefits:
    get:
      consumes:
      - application/json
      description: |-
        Получить список всех льгот с пагинацией, фильтрацией и умным поиском

        Поиск использует Full-Text Search MySQL с автоматическим поиском по частичному совпадению:
        - Поиск "транс" найдет "транспорт", "транспортный" и т.д.
        - Поиск "пенсион" найдет "пенсионер", "пенсионный" и т.д.
        - Каждое слово ищется с начала (prefix matching)

        Можно использовать Boolean операторы для сложных запросов:
        + обязательное слово: "+пенсионер +транспорт"
        - исключить слово: "льгота -студент"
        * явный wildcard: "транс*"
        "" точная фраза: "общественный транспорт"
      parameters:
      - description: Номер страницы (по умолчанию 1)
        in: query
        name: page
        type: integer
      - description: Количество элементов на странице (по умолчанию 10, максимум 100)
        in: query
        name: limit
        type: integer
      - description: ID региона для фильтрации
        in: query
        name: region
        type: integer
      - description: UUID города для фильтрации
        in: query
        name: city_id
        type: string
      - description: Тип льготы (federal, regional, commercial)
        in: query
        name: type
        type: string
      - description: Целевые группы через запятую (pensioners, disabled, students
          и т.д.)
        in: query
        name: target_groups
        type: string
      - description: Теги через запятую (most_popular, new, hot, best, recommended,
          popular, top)
        in: query
        name: tags
        type: string
      - description: Категории через запятую (medicine, transport, food, clothing,
          other)
        in: query
        name: categories
        type: string
      - description: Дата начала периода (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Дата окончания периода (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Поисковый запрос (автоматически ищет по частичному совпадению)
        in: query
        name: search
        type: string
      - description: Поле для сортировки (created_at, views, updated_at) - по умолчанию
          created_at
        in: query
        name: sort_by
        type: string
      - description: Направление сортировки (asc, desc) - по умолчанию desc
        in: query
        name: order
        type: string
      - description: Показать только избранные льготы (работает только при авторизации,
          иначе игнорируется)
        in: query
        name: favorites
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.benefitsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorStruct'
      summary: Get Benefits List
      tags:
      - Benefits
  /benefits/{id}:
    get:
      consumes:
      - application/json
      description: Получить льготу по ID
      parameters:
      - description: Benefit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.benefitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorStruct'
      summary: Get Benefit By ID
      tags:
      - Benefits
  /benefits/{id}/favorite:
    post:
      consumes:
      - application/json
      description: Переключить статус избранной льготы (toggle). Если льгота в избранном
        - удалит из избранного, если нет - добавит в избранное
      parameters:
      - description: Benefit ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorStruct'
      security:
      - UserAuth: []
      summary: Toggle Benefit Favorite Status
      tags:
      - Benefits
  /cities:
    get:
      consumes:
      - application/json
      description: Get all cities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.City'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorStruct'
      summary: Get Cities
      tags:
      - Cities
  /speech/recognize:
    post:
      consumes:
      - multipart/form-data
      description: Распознает речь из аудиофайла и возвращает текст
      parameters:
      - description: 'Аудиофайл (поддерживаются форматы: wav, mp3, ogg)'
        in: formData
        name: audio
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Распознанный текст
          schema:
            $ref: '#/definitions/v1.speechResponse'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Распознавание речи
      tags:
      - Speech
  /users/auth/callback:
    get:
      consumes:
      - application/json
      description: Callback endpoint для Auth - получает code от ESIA и редиректит
        на фронтенд
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to frontend with code
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
      summary: OAuth Callback
      tags:
      - Auth
  /users/auth/login:
    get:
      consumes:
      - application/json
      description: Перенаправление на ESIA для авторизации
      produces:
      - application/json
      responses:
        "302":
          description: Found
      summary: OAuth Login
      tags:
      - Auth
  /users/auth/token:
    post:
      consumes:
      - application/json
      description: Обмен authorization code на access и refresh токены (как в Keycloak)
      parameters:
      - description: Code и State
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.exchangeTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.exchangeTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
      summary: Exchange Code for Tokens
      tags:
      - Auth
  /users/pong:
    get:
      consumes:
      - application/json
      description: Pong
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
      security:
      - UserAuth: []
      summary: Pong
      tags:
      - Pong
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorStruct'
      security:
      - UserAuth: []
      summary: Get Profile
      tags:
      - Users
  /users/update-info:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: User update info request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userUpdateInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
      security:
      - UserAuth: []
      summary: User Update Info
      tags:
      - Users
swagger: "2.0"